- split service code
