+++ internal/app/file/file_repository.go
@@ -1,10 +1,14 @@
 package file
 
 import (
-	context "context"
+	"context"
 	"io"
 
 	"github.com/google/uuid"
+	apiRestaurantFile "github.com/kinneko-de/api-contract/golang/kinnekode/restaurant/file/v1"
+	"github.com/kinneko-de/restaurant-file-store-svc/internal/app/operation/logger"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 var (
@@ -14,4 +18,25 @@ var (
 type FileRepository interface {
 	CreateFile(ctx context.Context, fileId uuid.UUID, revisionId uuid.UUID) (io.WriteCloser, error)
 	OpenFile(ctx context.Context, fileId uuid.UUID, revisionId uuid.UUID) (io.ReadCloser, error)
+}
+
+func writeFile(stream apiRestaurantFile.FileService_StoreFileServer, fileId uuid.UUID, revisionId uuid.UUID) (uint64, []byte, error) {
+	fileWriter, err := FileRepositoryInstance.CreateFile(stream.Context(), fileId, revisionId)
+	if err != nil {
+		logger.Logger.Err(err).Msg("failed to create file")
+		return 0, nil, status.Error(codes.Internal, "failed to create file. please retry the request")
+	}
+
+	totalFileSize, sniff, err := receiveChunks(stream, fileWriter)
+	if err != nil {
+		return 0, nil, err
+	}
+
+	closeErr := fileWriter.Close()
+	if closeErr != nil {
+		logger.Logger.Err(closeErr).Msg("failed to close file")
+		return 0, nil, status.Error(codes.Internal, "failed to close file. please retry the request")
+	}
+
+	return totalFileSize, sniff, nil
 }
